Backend code refactorization
