+++ backend/src/graphql/inputs/createVisitor.input.ts
@@ -1,12 +1,13 @@
 import { InputType, Field } from '@nestjs/graphql';
 
 @InputType()
-export class UpdateVisitorInput {
-  @Field({ nullable: true })
-  email?: string;
+export class CreateVisitorInput {
 
-  @Field({ nullable: true })
-  password?: string;
+  @Field()
+  email: string;
+
+  @Field()
+  password: string;
 
   @Field({ nullable: true })
   visitor_fname?: string;
@@ -29,4 +30,9 @@ export class UpdateVisitorInput {
   @Field({ nullable: true })
   wed_venue?: string;
 
+  @Field({ nullable: true })
+  phone?: string;
+
+  @Field({ nullable: true })
+  city?: string;
 }
