--- backend/src/graphql/inputs/vendorFilter.ts
@@ -1,11 +1,0 @@
-import { InputType, Field  } from '@nestjs/graphql';
-
-@InputType()
-export class VendorFilterInput {
-  @Field({ nullable: true })
-  city?: string;
-
-  @Field({ nullable: true })
-  category?: string;
-
-}
